--- src/pages/PageA.js
@@ -1,10 +1,0 @@
-const PageA = () => {
-	return (
-		<div>
-			<h1>A</h1>
-			<p>a</p>
-		</div>
-	);
-};
-
-export default PageA;
