--- src/pages/PageB.js
@@ -1,10 +1,0 @@
-const PageB = () => {
-	return (
-		<div>
-			<h1>B</h1>
-			<p>b</p>
-		</div>
-	);
-};
-
-export default PageB;
