--- src/pages/PageC.js
@@ -1,10 +1,0 @@
-const PageC = () => {
-	return (
-		<div>
-			<h1>C</h1>
-			<p>c</p>
-		</div>
-	);
-};
-
-export default PageC;
